agent_hub_ci_bundle/requirements_ci.txt
# CI/CD dependencies for Agent Hub benchmarks
aiohttp>=3.8.0
matplotlib>=3.6.0
orjson>=3.8.0
PyYAML>=6.0
streamlit>=1.20.0
pandas>=1.5.0  # For visualization
numpy>=1.24.0  # For calculations